+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - Adds a `State` API, can be used to create an URL to share, see [State](documentation/STATE.md).
 - Fixed Bug where the title of a VcsTreeViewLeaf did overflow. Text Overflow is not set to ellipsis. The full title is shown in the html title
 - Adds a `VcsRadio` component for radio button groups.
+- Adds a `Legend` tool, where layer or style information can be shown, see [Legend](documentation/Legend.md).